+++ src/robot/entities/Roue.java
@@ -18,7 +18,7 @@ import robot.gui.BetterGUISimulator;
 public class Roue extends Robot{  
 	  
     public static final int volumeMaxRoue = 5000;
-    public static final double vitesseRemplissageRoue = 5000/600;
+    public static final int vitesseRemplissageRoue = 5000/600;
     public static final double vitesseDeversementRoue = 100/5;
     public static final double vitesseDeplacementRoue = 80 ;
     public static final Color couleurRoue = Color.yellow;
